reorg tests, cleanup
