+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/slo/index.ts
@@ -13,6 +13,7 @@ export default function ({ loadTestFile }: DeploymentAgnosticFtrProviderContext)
     loadTestFile(require.resolve('./delete_slo'));
     loadTestFile(require.resolve('./get_slo'));
     loadTestFile(require.resolve('./find_slo'));
+    loadTestFile(require.resolve('./find_slo_definition'));
     loadTestFile(require.resolve('./reset_slo'));
     loadTestFile(require.resolve('./update_slo'));
   });
